requirements.txt
customtkinter
pillow
numpy
exifread
matplotlib
rembg
opencv-python
